+++ src/cli/plugin/index.ts
@@ -1,15 +1,24 @@
 import type { CommandModule } from "yargs";
 import type yargs from "yargs";
+import { keygenCommand } from "./keygen";
 import { packCommand } from "./pack";
 import { infoCommand } from "./info";
+import { uploadCommand } from "./upload";
 import { setStability } from "../stability";
+import { initCommand } from "./init";
 
 const command = "plugin";
 
 const describe = "Commands for kintone plugin";
 
 const builder = (args: yargs.Argv) =>
-  args.command(infoCommand).command(packCommand).demandCommand();
+  args
+    .command(infoCommand)
+    .command(keygenCommand)
+    .command(packCommand)
+    .command(uploadCommand)
+    .command(initCommand)
+    .demandCommand();
 
 const handler = () => {
   /** noop **/